Access Permissions Added for All Users

+++ buses/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from . forms import BusForm
 
@@ -12,7 +13,11 @@ def bus_details(request, bus_no):
     bus = get_object_or_404(Buses, bus_no=bus_no)
     return render(request, 'bus_details.html', {'bus': bus})
 
+@login_required
 def bus_create(request):
+    if not request.user.profile.user_type in 'admin':
+        messages.error(request, 'You do not have admin permission to access this page.')
+        return redirect('/buses')  # Redirect to a relevant page
     if request.method == 'POST':
         form = BusForm(request.POST)
         if form.is_valid():
@@ -22,7 +27,11 @@ def bus_create(request):
         form = BusForm()
     return render(request, 'buses/bus_form.html', {'form': form})
 
+@login_required
 def bus_update(request, pk):
+    if not request.user.profile.user_type in 'admin':
+        messages.error(request, 'You do not have admin permission to access this page.')
+        return redirect('/buses')  # Redirect to a relevant page
     bus = get_object_or_404(Buses, pk=pk)
     if request.method == 'POST':
         form = BusForm(request.POST, instance=bus)
@@ -33,7 +42,11 @@ def bus_update(request, pk):
         form = BusForm(instance=bus)
     return render(request, 'buses/bus_form.html', {'form': form})
 
+@login_required
 def bus_delete(request, pk):
+    if not request.user.profile.user_type in 'admin':
+        messages.error(request, 'You do not have admin permission to access this page.')
+        return redirect('/buses')  # Redirect to a relevant page
     bus = get_object_or_404(Buses, pk=pk)
     if request.method == 'POST':
         bus.delete()
