+++ routes/views.py
@@ -3,6 +3,11 @@ from django.shortcuts import render, get_object_or_404, redirect
 from django.http import HttpResponse
 from .models import Route
 from .forms import RouteForm
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+
+
+
 
 def route_list(request):
     routes = Route.objects.all()
@@ -12,7 +17,11 @@ def route_detail(request, pk):
     route = get_object_or_404(Route, pk=pk)
     return render(request, 'routes/route_detail.html', {'route': route})
 
+@login_required
 def route_create(request):
+    if not request.user.profile.user_type in 'admin':
+        messages.error(request, 'You do not have admin permission to access this page.')
+        return redirect('/incharges')  # Redirect to a relevant page
     if request.method == 'POST':
         form = RouteForm(request.POST)
         if form.is_valid():
@@ -22,7 +31,11 @@ def route_create(request):
         form = RouteForm()
     return render(request, 'routes/route_form.html', {'form': form})
 
+@login_required
 def route_update(request, pk):
+    if not request.user.profile.user_type in 'admin':
+        messages.error(request, 'You do not have admin permission to access this page.')
+        return redirect('/incharges')  # Redirect to a relevant page
     route = get_object_or_404(Route, pk=pk)
     if request.method == 'POST':
         form = RouteForm(request.POST, instance=route)
@@ -33,7 +46,11 @@ def route_update(request, pk):
         form = RouteForm(instance=route)
     return render(request, 'routes/route_form.html', {'form': form})
 
+@login_required
 def route_delete(request, pk):
+    if not request.user.profile.user_type in 'admin':
+        messages.error(request, 'You do not have admin permission to access this page.')
+        return redirect('/incharges')  # Redirect to a relevant page
     route = get_object_or_404(Route, pk=pk)
     if request.method == 'POST':
         route.delete()
